Fix compilation on GCC.

+++ src/engine_lib/private/game/node/ui/LayoutUiNode.cpp
@@ -103,7 +103,7 @@ void LayoutUiNode::onMouseScrollMoveWhileHovered(int iOffset) {
     if (iOffset < 0) {
         iCurrentScrollOffset += std::abs(iOffset);
     } else if (iCurrentScrollOffset > 0) {
-        if (iOffset > iCurrentScrollOffset) {
+        if (static_cast<size_t>(iOffset) > iCurrentScrollOffset) {
             iCurrentScrollOffset = 0;
         } else {
             iCurrentScrollOffset -= iOffset;
@@ -242,7 +242,8 @@ void LayoutUiNode::recalculatePosAndSizeForDirectChildNodes() {
 
         // Add spacers to total portion sum.
         const auto spacerPortion = expandPortionSum * childNodeSpacing;
-        expandPortionSum += spacerPortion * (std::max(1ULL, mtxChildNodes.second.size()) - 1ULL);
+        expandPortionSum +=
+            spacerPortion * (std::max(static_cast<size_t>(1), mtxChildNodes.second.size()) - 1ULL);
 
         const auto spacerActualPortion = childExpandRule == ChildNodeExpandRule::DONT_EXPAND
                                              ? childNodeSpacing
